Version 0.5: now new rulesets are driven into the kernel automatically
optimizer now is looking for /root/auto-apply, if this file is
present _and_ executable, it calls

iptables-restore -c /root/auto-apply

So you can copy the file via slow lines and do a chmod +x later.
After applying the ruleset the file is renamed.

All the rest keeps untouched as it works well for now.